Move internal/dmrconst -> internal/dmr/dmrconst
