+++ internal/db/models/packet_test.go
@@ -23,7 +23,7 @@ import (
 	"testing"
 
 	"github.com/USA-RedDragon/DMRHub/internal/db/models"
-	"github.com/USA-RedDragon/DMRHub/internal/dmrconst"
+	"github.com/USA-RedDragon/DMRHub/internal/dmr/dmrconst"
 	"github.com/google/go-cmp/cmp"
 )
