+++ internal/db/models/user.go
@@ -24,7 +24,7 @@ import (
 	"time"
 
 	"github.com/USA-RedDragon/DMRHub/internal/config"
-	"github.com/USA-RedDragon/DMRHub/internal/dmrconst"
+	"github.com/USA-RedDragon/DMRHub/internal/dmr/dmrconst"
 	"github.com/USA-RedDragon/DMRHub/internal/http/api/utils"
 	gorm_seeder "github.com/kachit/gorm-seeder"
 	"gorm.io/gorm"
