+++ internal/dmr/calltracker/call_tracker.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/USA-RedDragon/DMRHub/internal/config"
 	"github.com/USA-RedDragon/DMRHub/internal/db/models"
-	dmrconst "github.com/USA-RedDragon/DMRHub/internal/dmrconst"
+	dmrconst "github.com/USA-RedDragon/DMRHub/internal/dmr/dmrconst"
 	"github.com/USA-RedDragon/DMRHub/internal/http/api/apimodels"
 	"github.com/USA-RedDragon/DMRHub/internal/logging"
 	"github.com/mitchellh/hashstructure/v2"
