+++ internal/dmr/dmrconst/const.go
